Move some small helper functions from objmodel.cpp to util.cpp
It's not too big of a move, hopefully this won't cause merge conflicts
for anyone. Just part of an attempt to reduce the size of objmodel.cpp
