+++ src/codegen/ast_interpreter.cpp
@@ -44,6 +44,7 @@
 #include "runtime/objmodel.h"
 #include "runtime/set.h"
 #include "runtime/types.h"
+#include "runtime/util.h"
 
 #ifndef NDEBUG
 #define DEBUG 1
