+++ src/runtime/inline/link_forcer.cpp
@@ -32,6 +32,7 @@
 #include "runtime/objmodel.h"
 #include "runtime/set.h"
 #include "runtime/types.h"
+#include "runtime/util.h"
 
 namespace pyston {
